backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# LLM
google-genai

# Database
sqlalchemy
psycopg2-binary
alembic

# Vector DB
pgvector

# Utilities
python-dotenv
pydantic
pydantic-settings
httpx

# Audio (Phase 3)
# openai
# python-multipart
